// client-clean-architecture/src/modules/fractal/presentation/types/stores/editor/scene-types.ts
export type DefaultScene = {
    sceneType: string;
    source: 'default';
};

export type PluginScene = {
    sceneType: string;
    source: 'plugin';
    analysisId: string;
    exposureId: string;
};

export type ColorCodingScene = {
    sceneType: string;
    source: 'color-coding';
    analysisId: string;
    exposureId: string;
    property: string;
    startValue: string;
    endValue: string;
    gradient: string;
};

export type ParticleFilterScene = {
    sceneType: 'particle-filter';
    source: 'particle-filter';
    analysisId?: string;
    exposureId?: string;
    property: string;
    operator: string;
    value: number;
    action?: string;
};

export type SceneObjectType = DefaultScene | PluginScene | ColorCodingScene | ParticleFilterScene;

export interface ModelData {
    modelBounds?: null,
    glbs: null
}

export type ActiveScene = SceneObjectType;

export interface ModelState {
    activeScene: SceneObjectType;
    activeScenes: ActiveScene[];
    activeModel: ModelData | null;
    isModelLoading: boolean;
    pointSizeMultiplier: number;
    sceneOpacities: Record<string, number>;
}

export interface ModelActions {
    selectModel: (glbs: any) => void;
    setGlbsWithoutLoading: (glbs: any) => void;
    resetModel: () => void;
    setIsModelLoading: (loading: boolean) => void;
    setModelBounds: (modelBounds: any) => void;
    setActiveScene: (scene: SceneObjectType) => void;
    addScene: (scene: SceneObjectType) => void;
    removeScene: (scene: SceneObjectType) => void;
    toggleScene: (scene: SceneObjectType) => void;
    setPointSizeMultiplier: (multiplier: number) => void;
    increasePointSize: () => void;
    decreasePointSize: () => void;
    setSceneOpacity: (sceneKey: string, opacity: number) => void;
    getSceneOpacity: (sceneKey: string) => number;
}

export type ModelStore = ModelActions & ModelState;